refactor: move telemetry configuration checks into configUtil class
Shares many imports with configUtil and configUtil can have one instance of ConfigAggregator
